+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Preserve the XLIFF `<source>` element when loading and dumping a file, instead of overwriting it with the message key
+ * Add a locale-aware `Plural-Forms` header to the output of `PoFileDumper`
 
 8.1
 ---